ADD) 에러핸들링 추가 및 baseurl 변경

See merge request s09-fintech-finance-sub2/S09P22A109!82

+++ BE/TikTagTalk/src/main/java/A109/TikTagTalk/domain/tagRoom/service/CommentServiceImpl.java
@@ -3,20 +3,25 @@ package A109.TikTagTalk.domain.tagRoom.service;
 import A109.TikTagTalk.domain.account.dto.response.ResponseDto;
 import A109.TikTagTalk.domain.account.dto.response.ResponseUtil;
 import A109.TikTagTalk.domain.account.entity.Account;
+import A109.TikTagTalk.domain.account.exception.NotExistException;
 import A109.TikTagTalk.domain.account.repository.AccountRepository;
 import A109.TikTagTalk.domain.tagRoom.dto.request.InsertCommentRequestDto;
 import A109.TikTagTalk.domain.tagRoom.dto.request.ModifyCommentRequestDto;
 import A109.TikTagTalk.domain.tagRoom.dto.response.AllCommentsResponseDto;
 import A109.TikTagTalk.domain.tagRoom.entity.Comment;
 import A109.TikTagTalk.domain.tagRoom.entity.TagRoom;
+import A109.TikTagTalk.domain.tagRoom.exception.CustomAccessDeniedException;
 import A109.TikTagTalk.domain.tagRoom.repository.CommentRepository;
 import A109.TikTagTalk.domain.tagRoom.repository.TagRoomRepository;
 import A109.TikTagTalk.domain.user.entity.Member;
 import A109.TikTagTalk.domain.user.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
+import org.apache.http.HttpStatus;
+import org.springframework.http.HttpStatusCode;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.nio.file.AccessDeniedException;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -53,7 +58,8 @@ public class CommentServiceImpl implements CommentService{
                     AllCommentsResponseDto.TagRoomDto tagRoomDto=AllCommentsResponseDto.TagRoomDto.builder()
                             .id(comment.getTagRoom().getId()).build();
                     AllCommentsResponseDto.MemberDto memberDto=AllCommentsResponseDto.MemberDto.builder()
-                            .id(comment.getMember().getId()).build();
+                            .id(comment.getMember().getId())
+                            .name(comment.getMember().getName()).build();
                     return AllCommentsResponseDto.builder()
                             .tagRoom(tagRoomDto)
                             .member(memberDto)
@@ -67,10 +73,10 @@ public class CommentServiceImpl implements CommentService{
     @Override
     @Transactional
     public ResponseDto modifyComment(ModifyCommentRequestDto requestDto, Member member) {
-        //현재 로그인된 member와 해당 comment 작성 member가 다르면 에러!!
+
         Member writer=commentRepository.findMemberByCommentId(requestDto.getId());
         if(member.getId()!=writer.getId()){
-            return ResponseUtil.Failure("권한이 없습니다.");
+            throw new CustomAccessDeniedException(HttpStatus.SC_FORBIDDEN,"권한이 없습니다.");
         }
         Comment comment=Comment.builder()
                 .id(requestDto.getId())
@@ -83,13 +89,17 @@ public class CommentServiceImpl implements CommentService{
     @Override
     @Transactional
     public ResponseDto deleteComment(Long commentId, Member member) {
+        if(!(commentRepository.findById(commentId).isPresent())){
+            throw new NotExistException(HttpStatus.SC_INTERNAL_SERVER_ERROR,"해당 comment을 찾을 수 없습니다.");
+        }
         Comment comment=commentRepository.findById(commentId).get();
+        System.out.println("commentRepository : "+commentRepository.findById(commentId)+"!!!");
         if(comment==null){
-            return ResponseUtil.Failure("댓글이 존재하지 않습니다.");
+            throw new NotExistException(HttpStatus.SC_INTERNAL_SERVER_ERROR,"해당 comment을 찾을 수 없습니다.");
         }
         Member writer=comment.getMember();
         if(member.getId()!=writer.getId()){
-            return ResponseUtil.Failure("권한이 없습니다.");
+            throw new CustomAccessDeniedException(HttpStatus.SC_FORBIDDEN,"권한이 없습니다.");
         }
         commentRepository.delete(comment);
         return ResponseUtil.Success("댓글 삭제 성공");
